Community-Supported/git-to-hyper/requirements.txt
click==8.1.3
GitPython==3.1.27
tableauhyperapi==0.0.15735